WIP refactor - making sense of how this code can be organized and extended
